+++ static/libs/code.js
@@ -1,1 +1,1 @@
-export{U as init}from"./code-BGIBNPzT.js";
+export{U as init}from"./code-jjdEvLco.js";