+++ static/libs/markdown.js
@@ -1,1 +1,1 @@
-export{w as init}from"./markdown-MoM2jSGX.js";
+export{w as init}from"./markdown-ZyPUwenx.js";